Comment code and fix spacing issues
